better error handaling and seperation of concerns and increased redability
